refactor medicos e pacientes:
-modo  de identificação foi mudado, usando cpf para pacientes e crm para médicos
-id interno ainda existente para relações de tabelas
-consultas, updates e deletes agora são feitos por cpf ou crm
-será introduzido também consultas por nomes

+++ app/controllers/medicos_controller.ts
@@ -2,6 +2,26 @@ import type { HttpContext } from '@adonisjs/core/http'
 import Medico from '#models/medicos'
 
 export default class MedicosController {
+  /**
+   * Creates a new Medico
+   *
+   * @body {Object} data
+   * @body {string} data.crm
+   * @body {string} data.nm_medico
+   * @body {string} data.nm_especializacao
+   *
+   * @response {Object} medico
+   * @response {number} medico.cd_medico
+   * @response {string} medico.crm
+   * @response {string} medico.nm_medico
+   * @response {string} medico.nm_especializacao
+   * @response {string} medico.created_at
+   * @response {string} medico.updated_at
+   *
+   * @status 201
+   * @throws {HttpException}
+   * @throws {ValidationError}
+   */
   async create({ request, response }: HttpContext) {
     try {
       const { data } = request.all()
@@ -18,9 +38,32 @@ export default class MedicosController {
     }
   }
 
+  /**
+   * Fetches a medico by its CRM
+   *
+   * @queryParam {string} crm required
+   *
+   * @response {Object} medico
+   * @response {number} medico.cd_medico
+   * @response {string} medico.crm
+   * @response {string} medico.nm_medico
+   * @response {string} medico.nm_especializacao
+   * @response {string} medico.created_at
+   * @response {string} medico.updated_at
+   *
+   * @status 200
+   * @throws {HttpException}
+   * @throws {ValidationError}
+   */
   async show({ params, response }: HttpContext) {
     try {
-      const medico = await Medico.findOrFail(params.id)
+      const cmr = params.crm
+
+      if (!cmr) {
+        return response.badRequest({ message: 'Medico ID is required' })
+      }
+
+      const medico = await Medico.findBy('crm', cmr)
 
       return response.ok(medico)
     } catch (error) {
@@ -28,13 +71,43 @@ export default class MedicosController {
     }
   }
 
+  /**
+   * Updates a medico by its CRM
+   *
+   * @queryParam {string} crm required
+   *
+   * @requestBody {Object} data
+   * @requestBody {number} data.cd_medico
+   * @requestBody {string} data.crm
+   * @requestBody {string} data.nm_medico
+   * @requestBody {string} data.nm_especializacao
+   *
+   * @response {Object} medico
+   * @response {number} medico.cd_medico
+   * @response {string} medico.crm
+   * @response {string} medico.nm_medico
+   * @response {string} medico.nm_especializacao
+   * @response {string} medico.created_at
+   * @response {string} medico.updated_at
+   *
+   * @status 200
+   * @throws {HttpException}
+   * @throws {ValidationError}
+   */
+
   async update({ params, request, response }: HttpContext) {
     try {
-      const medico = await Medico.findOrFail(params.id)
       const { data } = request.all()
+      const crm = params.crm
 
-      if (!data) {
+      if (!data || !crm) {
         return response.badRequest({ message: 'No data provided for update' })
+      }
+
+      const medico = await Medico.findBy('crm', crm)
+
+      if (!medico) {
+        return response.badRequest({ message: 'Medico not found' })
       }
 
       medico.merge(data)
@@ -46,9 +119,29 @@ export default class MedicosController {
     }
   }
 
+  /**
+   * Deletes a medico by its CRM
+   *
+   * @queryParam {string} crm required
+   *
+   * @status 204
+   * @throws {HttpException}
+   * @throws {ValidationError}
+   */
   async destroy({ params, response }: HttpContext) {
     try {
-      const medico = await Medico.findOrFail(params.id)
+      const crm = params.crm
+
+      if (!crm) {
+        return response.badRequest({ message: 'Medico ID is required' })
+      }
+
+      const medico = await Medico.findBy('crm', crm)
+
+      if (!medico) {
+        return response.notFound({ message: 'Medico not found' })
+      }
+
       await medico.delete()
       return response.noContent()
     } catch (error) {
